Finished working on avl tree deletion

+++ lab3/avl_tree.py
@@ -34,10 +34,6 @@ class AVL(BST):
                 node = current_node
             current_node = current_node.parent
         self.update_root()
-        if current_node == self.root:
-            return False
-        else:
-            return current_node
 
     def update_root(self):
         if self.root.parent is not None:
@@ -52,15 +48,54 @@ class AVL(BST):
         self.balance(new_node)
 
     def delete_node(self, number):
+        print(f"\nDELETED: {number}\n")
         node_to_delete = self.search(number)
-        node_to_balance = node_to_delete.parent
-        root_changed = node_to_delete.delete()
-        if root_changed:
-            self.root = root_changed
-        else:
-            next_node = self.balance(node_to_balance)
-            while next_node:
-                next_node = self.balance(next_node).parent
+        if self.root.left_child is None and self.root.right_child is None:
+            self.root = None
+            return
+        if self.root.left_child is None or self.root.right_child is None:
+            if self.root == node_to_delete:
+                root_changed = node_to_delete.delete()
+                self.root = root_changed
+            else:
+                node_to_delete.delete()
+            return
+        node_to_balance = node_to_delete.delete_AVL()
+        current_node = node_to_balance
+        while current_node is not None:
+            heights = current_node.get_child_heights()
+            current_node.height = 1 + max(heights)
+            balance = heights[1] - heights[0]
+            if balance == 2:
+                try:
+                    l_child_height = current_node.right_child.left_child.height
+                except AttributeError:
+                    l_child_height = 0
+                try:
+                    r_child_height = current_node.right_child.right_child.height
+                except AttributeError:
+                    r_child_height = 0
+                if l_child_height <= r_child_height:
+                    current_node.rotate_left()
+                else:
+                    current_node.right_child.rotate_right()
+                    current_node.rotate_left()
+            if balance == -2:
+                try:
+                    l_child_height = current_node.left_child.left_child.height
+                except AttributeError:
+                    l_child_height = 0
+                try:
+                    r_child_height = current_node.left_child.right_child.height
+                except AttributeError:
+                    r_child_height = 0
+                if l_child_height <= r_child_height:
+                    current_node.left_child.rotate_left()
+                    current_node.rotate_right()
+                else:
+                    current_node.rotate_right()
+            current_node = current_node.parent
+        self.update_root()
 
 
 class Node_AVL(Node):
@@ -115,14 +150,26 @@ class Node_AVL(Node):
                 heights.append(child.height)
         return heights
 
+    def delete_AVL(self):
+        i_pred = super().delete()
+        if i_pred is not None:
+            return i_pred.parent
+        if self.right_child and self.left_child:
+            return self
+        if self.left_child:
+            return self.left_child
+        if self.right_child:
+            return self.right_child
+        return self.parent
+
 
 if __name__ == "__main__":
-    # numbers = sample(range(0, 15), 10)
-    # print(numbers)
-    # deleted = sample(numbers, 3)
-    # print(deleted)
-    numbers = [12, 1, 6, 14, 0, 11, 7, 10, 8, 4]
-    deleted = [4, 0, 6]
+    numbers = sample(range(0, 150), 50)
+    print(numbers)
+    deleted = sample(numbers, 50)
+    print(deleted)
+    # numbers = [12, 1, 6, 14, 0, 11, 7, 10, 8, 4]
+    # deleted = [4, 0, 6, 12, 10, 11, 14, 8, 1,7]
     tree = AVL(numbers)
     # print(tree)
     for num in deleted:
